Support sigwinch on Windows with huge changes
