+++ option.go
@@ -1,5 +1,7 @@
 package prompt
 
+import "fmt"
+
 // Option is the type to replace default parameters.
 // prompt.New accepts any number of options (this is functional option pattern).
 type Option func(prompt *Prompt) error
@@ -268,7 +270,10 @@ func OptionAddASCIICodeBind(b ...ASCIICodeBind) Option {
 
 // New returns a Prompt with powerful auto-completion.
 func New(executor Executor, completer Completer, opts ...Option) *Prompt {
-	defaultWriter := NewStandardOutputWriter()
+	defaultWriter, err := NewStandardOutputWriter()
+	if err != nil {
+		panic(fmt.Sprintf("new stdout writer: %s", err))
+	}
 	registerConsoleWriter(defaultWriter)
 
 	pt := &Prompt{
